add live capture and halton to generate evenly distributed points in a 3d space

+++ src/types.rs
@@ -12,5 +12,5 @@ pub struct AppSettings {
     pub model_reposition_sensitivity: f32,
     pub mouse_sensitivity: f32,
     pub zoom_sensitivity: f32,
-    pub live_capture_iteration: u32,
+    pub live_capture_iteration: usize,
 }
